Получение TODO на дату

+++ main.py
@@ -12,7 +12,7 @@ bot = telebot.TeleBot(API_TOKEN)
 
 users = {}
 users_todo = {}
-
+todo_from_file = list()
 
 def is_valid_name_surname(name_surname):
     return not (" " in name_surname or len(name_surname) < 2)
@@ -27,14 +27,53 @@ def start(message):
         remove_initial_keyboard(user_id, "Как тебя зовут?")
         bot.register_next_step_handler(message, get_name)
     elif message.text == "TODO":
-        users_todo["user_id"] = user_id
+        users_todo[user_id] = {}
         remove_initial_keyboard(user_id, "Введите текст заметки")
         bot.register_next_step_handler(message, get_text_todo)
 
-    # elif message.text == "Test":
-    #     remove_initial_keyboard(user_id, "Test")
+    elif message.text == "Get TODO":
+        csv_dir = os.path.join("TODOInformation", "csv")
+        file_path = os.path.join(csv_dir, "todos.csv")
+        todo_from_file.clear()
+
+        with open(file_path) as csv_file:
+            reader = csv.DictReader(csv_file)
+            for csv_str in reader:
+                todo_from_file.append(csv_str)
+
+        remove_initial_keyboard(user_id, "Введите дату заметки")
+        bot.register_next_step_handler(message, get_text_todo_from_file)
+
     else:
         render_initial_keyboard(user_id)
+
+def get_text_todo_from_file(messege):
+    user_id = messege.from_user.id
+    if todo_from_file:
+        try:
+            text_date_todo = messege.text.title()
+            date_todo = datetime.strptime(text_date_todo, "%d/%m/%Y")
+            required_data = list(find_all_by_key(todo_from_file, "user_id", "date", str(user_id), text_date_todo))
+
+            if required_data:
+                bot.send_message(user_id, f"На дату {text_date_todo} найдены следующие заметки:")
+
+                counter = 1
+                for required in required_data:
+                    required_str = required["text"]
+
+                    bot.send_message(user_id, f"{counter}. {required_str}")
+                    counter +=1
+            else:
+                bot.send_message(user_id, f"На дату {text_date_todo} нет заметок")
+
+        except ValueError:
+            bot.send_message(user_id, "Введи дату корректно!")
+            bot.register_next_step_handler(messege, get_text_todo_from_file)
+
+def find_all_by_key(iterable, key1, key2, value1, value2):
+    return (dict_ for dict_ in iterable
+            if dict_[key1] == value1 and dict_[key2] == value2)
 
 def get_text_todo(message):
     user_id = message.from_user.id
@@ -42,7 +81,7 @@ def get_text_todo(message):
     # bot.send_message(user_id, f"{text_todo}")
 
     if text_todo:
-        users_todo["text"] = text_todo
+        users_todo[user_id]["text"] = text_todo
 
         bot.send_message(user_id, "Введи дату заметки в формате dd/mm/yyyy")
         bot.register_next_step_handler(message, get_date_todo)
@@ -58,13 +97,13 @@ def get_date_todo(messege):
 
     try:
         date_todo = datetime.strptime(text_date_todo, "%d/%m/%Y")
-        users_todo["date"] = text_date_todo
+        users_todo[user_id]["date"] = text_date_todo
 
-        todo_text = users_todo["text"]
-        todo_date = users_todo["date"]
+        todo_text = users_todo[user_id]["text"]
+        todo_date = users_todo[user_id]["date"]
         question = f"Ты создал заметку с текстом \"{todo_text}\" на дату {todo_date}. Верно?"
         render_yes_now_keyboard(user_id, question, "to")
-    except:
+    except ValueError:
         bot.send_message(user_id, "Введи дату корректно!")
         bot.register_next_step_handler(messege, get_date_todo)
 
@@ -161,15 +200,24 @@ def callback_worker(call):
 
         json_dir = os.path.join("TODOInformation", "csv")
         file_path = os.path.join(json_dir, "todos.csv")
+
+        fieldnames = ["text", "date"]
+        fieldmap = dict(zip(fieldnames, fieldnames))
+
         if not os.path.exists(json_dir):
             os.makedirs(json_dir)
 
-        with open(file_path, "w") as csv_file:
-            fieldnames = ["user_id", "text", "date"]
-            writer = csv.DictWriter(csv_file, fieldnames=fieldnames)
+        if not os.path.isfile(file_path):
+            with open(file_path, "w") as csv_file:
+                writer = csv.DictWriter(csv_file, fieldnames=["user_id"] + fieldnames)
+                writer.writeheader()
 
-            writer.writeheader()
-            writer.writerow(users_todo)
+        with open(file_path, "a") as csv_file:
+            writer = csv.DictWriter(csv_file, fieldnames=["user_id"] + fieldnames)
+
+            for user_id, date_values in users_todo.items():
+                row = {fieldname: date_values[fieldmap[fieldname]] for fieldname in fieldnames}
+                writer.writerow(dict(user_id=user_id, **row))
 
         bot.send_message(user_id, "Готово!")
 
@@ -192,8 +240,9 @@ def render_initial_keyboard(user_id: int):
     keyboard = types.ReplyKeyboardMarkup(row_width=2, resize_keyboard=True)
     register_button = types.KeyboardButton("Регистрация")
     todo_button = types.KeyboardButton("TODO")
+    get_todo_button = types.KeyboardButton("Get TODO")
     # test_button = types.KeyboardButton("Test")
-    keyboard.add(register_button, todo_button)
+    keyboard.add(register_button, todo_button, get_todo_button)
     bot.send_message(user_id, "Выберите действие", reply_markup=keyboard)
